+++ lib/tests/test_completions.rs
@@ -10,6 +10,11 @@ fn test_completions() -> Result<()> {
     // let help_text = run_xvc(None, &[], XvcVerbosity::Default)?;
     // assert!(!help_text.is_empty());
 
+    // Nushell only has static completions for the time.
+    let nushell_completions = run_xvc(None, &["_comp", "generate-nushell"], XvcVerbosity::Default)?;
+
+    assert!(!nushell_completions.is_empty());
+
     let shells = ["bash", "elvish", "fish", "powershell", "zsh"];
 
     for s in shells {